+++ bpf/libbpf/libbpf_wrapper.h
@@ -9,3 +9,7 @@ struct perf_buffer *tl_perf_buffer__new(int map_fd,
 					size_t page_cnt,
 					perf_buffer_sample_fn sample_cb,
 					perf_buffer_lost_fn lost_cb);
+
+int tl_bpf_map_update_elem(const struct bpf_map *map, const void *key,
+			   size_t key_sz, const void *value, size_t value_sz,
+			   __u64 flags);