+++ bpf/libbpf/spin_lock.h
@@ -1,0 +1,4 @@
+struct spin_lock_hmap_elem {
+	struct bpf_spin_lock lock;
+	int var[2];
+};