+++ bpf/libbpf/tc.c
@@ -17,7 +17,14 @@
 #include <linux/bpf.h>
 #include "btf_helpers.h"
 #include "trace_helpers.h"
-#if defined(TEST_RBTREE)
+#if defined(TEST_SPIN_LOCK)
+#include "libbpf_wrapper.h"
+#include "spin_lock.h"
+#include "spin_lock.skel.h"
+#define struct_bpf	spin_lock_bpf
+#define _bpf__open_and_load	spin_lock_bpf__open_and_load
+#define _bpf__destroy	spin_lock_bpf__destroy
+#elif defined(TEST_RBTREE)
 #include "rbtree.skel.h"
 #define struct_bpf	rbtree_bpf
 #define _bpf__open_and_load	rbtree_bpf__open_and_load
@@ -111,6 +118,21 @@ int main(int argc, char **argv)
 		fprintf(stderr, "Failed to open BPF skeleton\n");
 		return 1;
 	}
+
+#if defined(TEST_SPIN_LOCK)
+	printf("spin_lock hash map max entries %d\n",
+		bpf_map__max_entries(skel->maps.spin_lock_hash_map));
+
+	int zero = 0;
+	struct spin_lock_hmap_elem elem = {0};
+	err = tl_bpf_map_update_elem(skel->maps.spin_lock_hash_map,
+				&zero, sizeof(zero),
+				&elem, sizeof(elem), 0);
+	if (err) {
+		fprintf(stderr, "Failed to update hash map: %d\n", err);
+		goto cleanup;
+	}
+#endif
 
 	/**
 	 * The hook (i.e. qdisc) may already exists because: